Secure the API and fix up permissions
The API key was previously set as an attribute meaning that anyone
with access to the cookbook repo could see and use the API key.  This
change moves the API Key into an Encrypted Data Bag.

The permissions on /opt/dataloop were constantly being reset to
"root:root" preventing the agent from starting properly.  This change
creates the "dataloop" user, sets the home directory to
"/opt/dataloop" and runs a small bash script over that directory to
ensure that the permissions are correct.